Improved Theoretical Analysis
It's still not final, but it's more clear now (the point about 2log(N+1) auxiliary space for LLRB BST). Also, I hope we can clearly distinguish what goes here and what goes for experimental analysis, especially with Bloom Filter where the line is kinda blurry. I mean the choice of the size of the bitarray as part of minimising false positives.